launcher/requirements.txt
pystray>=0.2.8
Pillow>=10.0.0
